updated script

+++ iso/archiso/airootfs/root/customize_airootfs.sh
@@ -18,18 +18,26 @@ EOF
 
 echo ">>> Setting xfce4-terminal for Nemo..."
 
-# Start a temp D-Bus session so gsettings can work
-eval $(dbus-launch)
+if command -v gsettings >/dev/null 2>&1; then
+  eval $(dbus-launch)
 
-# Set terminal preference for Nemo
-gsettings set org.cinnamon.desktop.default-applications.terminal exec "xfce4-terminal"
-gsettings set org.cinnamon.desktop.default-applications.terminal exec-arg "-e"
+  gsettings set org.cinnamon.desktop.default-applications.terminal exec "xfce4-terminal"
+  gsettings set org.cinnamon.desktop.default-applications.terminal exec-arg "-e"
 
-# Copy Dconf settings to /etc/skel for installed user
-mkdir -p /etc/skel/.config/dconf
-cp /root/.config/dconf/user /etc/skel/.config/dconf/user
+  mkdir -p /etc/skel/.config/dconf
+  cp /root/.config/dconf/user /etc/skel/.config/dconf/user
 
-# Clean up D-Bus
-kill $DBUS_SESSION_BUS_PID
-unset DBUS_SESSION_BUS_PID
+  kill $DBUS_SESSION_BUS_PID
+  unset DBUS_SESSION_BUS_PID
+fi
+
+## Spotify adblock patch
+sed -i 's|^Exec=spotify|Exec=env LD_PRELOAD=/usr/lib/spotify-adblock.so spotify %U|' /usr/share/applications/spotify.desktop
+
+
+### To Fish We Go
+echo "Setting fish as default shell system-wide..."
+grep -qxF '/usr/bin/fish' /etc/shells || echo '/usr/bin/fish' >> /etc/shells
+chsh -s /usr/bin/fish root
+sed -i 's|^SHELL=.*|SHELL=/usr/bin/fish|' /etc/default/useradd
